+++ mail/mail-uniffi/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2024-00-00
 
+### Added
+
+- Added `watch_available_move_to_actions`.
+
 ## [0.21.3] - 2024-11-08
 
 ### Fix