+++ mail/mail-uniffi/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2024-00-00
 
+## [0.21.4] - 2024-11-13
+
 ### Added
 
   - Removed InAppPromosHidden field from API